feat(layout): :lipstick: minor revamp (#47)
switch font, add more whitespace
