+++ package.json
@@ -21,8 +21,7 @@
     "@chakra-ui/react": "^1.8.1",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
-    "@fontsource/inter": "^4.5.2",
-    "@fontsource/jost": "^4.5.2",
+    "@fontsource/outfit": "^4.5.3",
     "firebase": "^9.6.5",
     "formik": "^2.2.9",
     "framer-motion": "^5.6.0",
