+++ src/lib/styles/theme/components/button.ts
@@ -1,0 +1,10 @@
+import type { DeepPartial, Theme } from "@chakra-ui/react";
+
+export const Button: DeepPartial<Theme["components"]["Button"]> = {
+  baseStyle: {
+    borderRadius: 12,
+  },
+  defaultProps: {
+    size: "lg",
+  },
+};
