+++ src/lib/styles/theme/components/index.ts
@@ -1,0 +1,7 @@
+import type { DeepPartial, Theme } from "@chakra-ui/react";
+
+import { Button } from "./button";
+
+export const components: DeepPartial<Theme["components"]> = {
+  Button,
+};
