+++ src/lib/styles/theme/fonts.ts
@@ -1,0 +1,6 @@
+import type { DeepPartial, Theme } from "@chakra-ui/react";
+
+export const fonts: DeepPartial<Theme["fonts"]> = {
+  heading: "Outfit, sans-serif",
+  body: "Outfit, sans-serif",
+};
