+++ src/lib/styles/theme/index.ts
@@ -1,0 +1,11 @@
+import { extendTheme } from "@chakra-ui/react";
+
+import { components } from "./components";
+import { fonts } from "./fonts";
+
+const customTheme = extendTheme({
+  fonts,
+  components,
+});
+
+export default customTheme;
