+++ src/pages/intro.tsx
@@ -1,7 +1,6 @@
-import { Button, Container, Heading, Text } from "@chakra-ui/react";
+import { Box, Button, Container, Heading, Text } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 
-import SomeInfo from "lib/components/hall/components/SomeInfo";
 import SpokerWrapperGrid from "lib/components/ui/SpokerWrapperGrid";
 import { trackEvent } from "lib/utils/trackEvent";
 
@@ -14,19 +13,26 @@ const Intro = () => {
   };
 
   return (
-    <Container paddingX={0} display="grid" gridGap={8}>
-      <SpokerWrapperGrid textAlign="center" gap={8}>
-        <Heading>Estimate with your team</Heading>
+    <Container
+      paddingX={0}
+      display="grid"
+      gridGap={8}
+      minHeight="60vh"
+      alignItems="center"
+    >
+      <Box>
+        <SpokerWrapperGrid textAlign="center" gap={8}>
+          <Heading>Estimate with your team</Heading>
 
-        <Text>
-          Welcome to my take on scrum poker web app. Just click start below.
-        </Text>
+          <Text>
+            Welcome to my take on scrum poker web app. Just click start below.
+          </Text>
 
-        <Button colorScheme="pink" onClick={handleClickStart}>
-          Start
-        </Button>
-      </SpokerWrapperGrid>
-      <SomeInfo />
+          <Button colorScheme="pink" onClick={handleClickStart}>
+            Start
+          </Button>
+        </SpokerWrapperGrid>
+      </Box>
     </Container>
   );
 };
